+++ src/Creature.hpp
@@ -17,7 +17,7 @@ class Creature : public Entity{
 public:
     int health;
     float movementSpeed;
-    Creature(int x, int y, float movementSpeed, int health, bool visible, std::string imgPath);
+    Creature(float x, float y, float movementSpeed, int health, bool visible, std::string imgPath);
     int getHealth();
     void takeDamage(int dmg);
     void moveTo(float x, float y);
